--- a/modules/apps/journal/journal-test/src/testIntegration/resources/com/liferay/journal/service/test/dependencies/complex_data_definition.json
+++ b/modules/apps/journal/journal-test/src/testIntegration/resources/com/liferay/journal/service/test/dependencies/complex_data_definition.json
@@ -1,6 +1,18 @@
 {
 	"availableLanguageIds": [
-		"es_ES"
+		"ar_SA",
+		"ca_ES",
+		"de_DE",
+		"en_US",
+		"es_ES",
+		"fi_FI",
+		"fr_FR",
+		"hu_HU",
+		"ja_JP",
+		"nl_NL",
+		"pt_BR",
+		"sv_SE",
+		"zh_CN"
 	],
 	"contentType": "journal",
 	"dataDefinitionFields": [