--- a/modules/apps/journal/journal-test/src/testIntegration/resources/com/liferay/journal/service/test/dependencies/complex_template.ftl
+++ b/modules/apps/journal/journal-test/src/testIntegration/resources/com/liferay/journal/service/test/dependencies/complex_template.ftl
@@ -1,0 +1,9 @@
+<p>Web Content Render</p>
+
+<#if repeatableImage.getSiblings()?has_content>
+	<#list repeatableImage.getSiblings() as cur_repeatableImage>
+		<#if (cur_repeatableImage.getData())?? && cur_repeatableImage.getData() != "">
+			<img alt="${cur_repeatableImage.getAttribute("alt")}" data-fileentryid="${cur_repeatableImage.getAttribute("fileEntryId")}" src="${cur_repeatableImage.getData()}" />
+		</#if>
+	</#list>
+</#if>