… modules/apps/journal/journal-test/src/testIntegration/resources/com/liferay/journal/service/test/dependencies/complex_journal_content.xml
… modules/apps/journal/journal-test/src/testIntegration/resources/com/liferay/journal/service/test/dependencies/complex_with_translations_journal_content.xml