https://liferay.atlassian.net/browse/LPS-194835

Hey @ealonso ,

As mentioned, this PR:
- Modifies **testGetArticleDisplayWithComplexData** to: Add an additional 100 fields (1 translation per field)
- Adds **testGetArticleDisplayWithComplexDataAndTranslations** to: Add multiple translations (13 translations per field) (which also results in additional fields)

With these two tests, from my manual testing, we can observe the performance improvement from [LPS-184283](https://liferay.atlassian.net/browse/LPS-184283)
- Before fix:
  - testGetArticleDisplayWithComplexData in ~341 ms
  - testGetArticleDisplayWithComplexDataAndTranslations in ~670 ms
- After fix:
  - testGetArticleDisplayWithComplexData in ~277 ms
  - testGetArticleDisplayWithComplexDataAndTranslations in ~298 ms

**Concern:** required adding an additional: `complex_with_translations_journal_content.xml`. 
- Could be difficult/inefficient to maintain both **complex_journal_content.xml** and **complex_with_translations_journal_content.xml**

Please let me know if you have any questions.
Thanks!