@liferay-continuous-integration 2023-08-29
<h1>Build started.</h1><p>Jenkins is currently running tests in the <code>sf</code> test suite.</p><h4>Base Branch:</h4><p>Branch Name: <a href="https://github.com/liferay/liferay-portal/tree/master">master</a></p><h4>Job Summary:</h4><p>Job Link: <a href="https://test-1-16.liferay.com/job/test-portal-source-format/6711/">test-portal-source-format</a></p><p><b>For more details click <a href="https://test-1-16.liferay.com/userContent/jobs/test-portal-source-format/builds/6711/jenkins-report.html">here</a>.</b></p>