@liferay-continuous-integration 2023-08-29
<html><h3>:heavy_check_mark: ci:test:sf - 1 out of 1 jobs passed in 9 minutes</h3><details><summary>Click here for more details.</summary><h4>Base Branch:</h4><p>Branch Name: <a href="https://github.com/liferay/liferay-portal/tree/master">master</a><br/>Branch GIT ID: <a href="https://github.com/liferay/liferay-portal/commit/b868e9a79f3de7975f6b5a24c8be94f689e9eeff">b868e9a79f3de7975f6b5a24c8be94f689e9eeff</a></p><h4>Sender Branch:</h4><p>Branch Name: <a href="https://github.com/ericyanLr/liferay-portal/tree/LPS-194835">LPS-194835</a><br/>Branch GIT ID: <a href="https://github.com/ericyanLr/liferay-portal/commit/9ed4a45ed8bf00785aa1c10ca71f3039bdcc6478">9ed4a45ed8bf00785aa1c10ca71f3039bdcc6478</a></p>1 out of 1jobs PASSED<details><summary><strong>1 Successful Jobs:</strong></summary><ul><li><a href="https://test-1-16.liferay.com/job/test-portal-source-format/6711/">test-portal-source-format</a></li></ul></details><h5>For more details click <a href="https://test-1-16.liferay.com/userContent/jobs/test-portal-source-format/builds/6711/jenkins-report.html">here</a>.</h5></details></html>